Refactor deriva-py project structure into a single package repo. Remove deriva_io as a subpackage and create new repo for it.
