--- deriva_io/docs/README.md
@@ -1,2 +1,0 @@
-# deriva_io
-IO modules for Python-based DERIVA platform code.
