+++ deriva_py/common/__init__.py
@@ -11,8 +11,6 @@ from collections import OrderedDict
 from distutils.util import strtobool
 from portalocker import Lock, LOCK_EX, LOCK_SH
 
-__version__ = "0.2.2"
-
 if sys.version_info > (3,):
     from urllib.parse import quote as urlquote
 else:
@@ -180,7 +178,4 @@ def resource_path(relative_path, default=os.path.abspath(".")):
         return relative_path
     return os.path.join(default, relative_path)
 
-from .hatrac_store import HatracStore, HatracHashMismatch, HatracJobPaused, HatracJobAborted
-from .ermrest_catalog import ErmrestCatalog
-from .ermrest_config import AttrDict, CatalogConfig
-from .polling_ermrest_catalog import PollingErmrestCatalog
+
